requirements.txt
aiogram==3.4.1
asyncpg==0.29.0
python-dotenv==1.0.0
aiohttp==3.9.1
pydantic==2.5.0
pydantic-settings==2.1.0
gigachat==0.1.3